+++ server/standalone/dbal/computeplan.go
@@ -100,7 +100,7 @@ from "compute_plans" as cp
 left join "compute_tasks" as t on (t.asset->>'computePlanKey')::uuid = cp.id and t.channel = cp.channel
 where cp.channel=$3
 group by cp.asset, cp.asset->>'creationDate'
-order by cp.asset->>'creationDate' asc limit $1 offset $2
+order by cp.asset->>'creationDate' asc, id limit $1 offset $2
 `
 
 	rows, err = d.tx.Query(d.ctx, query, p.Size+1, offset, d.channel)