+++ server/standalone/dbal/computetask.go
@@ -254,7 +254,7 @@ func (d *DBAL) QueryComputeTasks(p *common.Pagination, filter *asset.TaskQueryFi
 	builder := pgDialect.Select("asset").
 		From("compute_tasks").
 		Where(squirrel.Eq{"channel": d.channel}).
-		OrderByClause("asset->>'creationDate' ASC").
+		OrderByClause("asset->>'creationDate' ASC, id").
 		Offset(uint64(offset)).
 		Limit(uint64(p.Size + 1))