+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Stable sorting of tasks (#371)
+
 ## [0.0.2] - 2021-09-16
 
 ### Added